// semFront/src/app/models/promocionTicket-interface.ts
export interface PromocionTicketInterface{
    id?: string;
    titulo?: string;
    numVentas?: any;
    descripcion?: string;
    indicaciones?: string;
    precioVenta?: string;
    imagen?: string;
    estudioInicial?: string;
    caducidad?: any;
    Catalogo?: any;
    sexo ?: any;
    results ?: any;
    count ?: any;
}