// semFront/src/app/models/cita-interface.ts
export interface CitaInterface{
    id?: string;
    fecha_cita?: string;
    hora_inicio?: string;
    hora_final?: string;
    tipo_cita?: string;
    id_sala?: string;
    prueba?: string;
    categoria?: string;
    precioVenta?: any;
    costoSucursal?: any;
    notas?: string;
    razon_rembolso?: string;
    estatus_rembolso?: string;
    estatus?: string;
    creador?: string;
    editor?: string;
    id_pago?: string;
    promocion?: string;
    nombre_promocion?: string;
    tipo_pago?: string;
    app_movil?: string;
    otro_nombre_paciente?: string;
    otro_fecha_nacimento?: string;
    otro_parentesco_paciente?: string;
    creacion?: string;
    ultimaActualizacion?: string;
    count?: any;
    results?: any;
    Sucursal?: any;
    Paciente?: any;
}